requirements.txt
# Core LLM stack
langchain
langchain-community
langchain-openai
openai

# Embeddings
sentence-transformers==2.2.2
transformers==4.30.2

# Vector DB
chromadb

# PDF parsing
PyMuPDF
pdfminer.six

# Optional: used by some loaders
unstructured

# Streamlit UI
streamlit

# Env management
python-dotenv

#
